Refactor: type 패키지명 변경
