+++ product-management-api/src/test/java/org/ecommerce/productmanagementapi/service/ProductManagementServiceTest.java
@@ -11,10 +11,10 @@ import java.util.Optional;
 import org.ecommerce.common.error.CustomException;
 import org.ecommerce.product.entity.Product;
 import org.ecommerce.product.entity.SellerRep;
-import org.ecommerce.product.entity.type.Acidity;
-import org.ecommerce.product.entity.type.Bean;
-import org.ecommerce.product.entity.type.ProductCategory;
-import org.ecommerce.product.entity.type.ProductStatus;
+import org.ecommerce.product.entity.enumerated.Acidity;
+import org.ecommerce.product.entity.enumerated.Bean;
+import org.ecommerce.product.entity.enumerated.ProductCategory;
+import org.ecommerce.product.entity.enumerated.ProductStatus;
 import org.ecommerce.productmanagementapi.dto.ProductManagementDto;
 import org.ecommerce.productmanagementapi.exception.ProductManagementErrorCode;
 import org.ecommerce.productmanagementapi.repository.ImageRepository;
