+++ product-search-api/src/main/java/org/ecommerce/productsearchapi/dto/ProductSearchDto.java
@@ -3,10 +3,10 @@ package org.ecommerce.productsearchapi.dto;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import org.ecommerce.product.entity.type.Acidity;
-import org.ecommerce.product.entity.type.Bean;
-import org.ecommerce.product.entity.type.ProductCategory;
-import org.ecommerce.product.entity.type.ProductStatus;
+import org.ecommerce.product.entity.enumerated.Acidity;
+import org.ecommerce.product.entity.enumerated.Bean;
+import org.ecommerce.product.entity.enumerated.ProductCategory;
+import org.ecommerce.product.entity.enumerated.ProductStatus;
 
 import com.google.common.annotations.VisibleForTesting;
